Dockerfile unified, both ports reachable
It should be noted that we cannot use a scratch container, as we have to
use `sh` to run the server startup command and `chmod` executable
permissions. Additionally, the `server` pkg has an error in the
`server/slack/connect.go` file that was panicking and preventing the
container from starting either `server` or `zendesk`.
I created a debug environment to test, and both ports were reachable in
that environment.